COM-171 Update function drugIsARV to lookup by drug name
The previous version of this function verified if a drug is ARV by checking if the drug id is within a given range. Some of the drugs within that range are not ARVs. This commit fixes this issue by verifying if a drug name is part of the given list of ARVs