COM-88 Remove unnecessary check
By enforcing the condition that the child has to be registered before the reporting period start date, we are excluding the children registered during the reporting period that have a viral test done during the period. As these children won't be counted for the next quarter.

The check is not necessary as the child needs to be registered to be able to have a test done.